+++ pkg/confkafka/message.go
@@ -1,0 +1,5 @@
+package confkafka
+
+import (
+	_ "github.com/segmentio/kafka-go"
+)