+++ pkg/confpulsar/extra_key.go
@@ -1,0 +1,8 @@
+package confpulsar
+
+// ExtraKey defines keys for extended metadata for extending more mq-specific features
+type ExtraKey string
+
+const (
+	EXTRA_KEY__EXPIRED_AT = "EXPIRED_AT"
+)