+++ pkg/confrabbitmq/message.go
@@ -1,0 +1,6 @@
+package confrabbitmq
+
+import (
+	_ "github.com/rabbitmq/amqp091-go"
+	_ "github.com/wagslane/go-rabbitmq"
+)